fix circular dependency warning
Circular dependency warning when building with Rollup #177
